+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Fixing version
+- Refactor dashboard server UI